Create README.md
In a well saturated and diverse market, cosmetologists face the challenge of drawing in new business on an almost daily basis. This is why I’d like to develop a modernized and responsive salon booking website for my wife as she continues to work independently. The goal is to give her current and future clients the ability to book appointments with her from their mobile devices as well as give my wife the ability to take full control of her daily schedule. This website will also utilize her current Square account. This will generate new appointments and sync them directly to her calendar while also serving as her point-of-sale system and notification system for client appointment times. This website will stay in development until it reaches industry standards.